requirements.txt
Flask==3.0.3
gunicorn==22.0.0
python-dotenv==1.0.1
requests==2.32.3
google-cloud-firestore==2.16.1
pypdf==4.1.0
openai>=1.40.0
httpx>=0.27.0
google-cloud-tasks